requirements.txt
flask
line-bot-sdk==3.4.0
sqlalchemy
psycopg2-binary
apscheduler
pytz